Win basepath fix & avidemo and renderer additions
Fixed the basepath problem on windows, now the users files will stay on
the games folder structure and only linux version is allowed to use the
users home path. Added jpeg cl_avidemo via cl_avidemotype cvar and upped
the avidemo image limit from 9999 to 99999. Added zbuffer writer to the
render for the future implementation of the Depth of field effect.