ui: reverted to better credits pages usuability
This reverts to the old 'fashion' credits pages usuability. Navigation
between the pages are done with virtual "next" and "previous" style
buttons. A single click on the 'BACK' button returns to the main menu
(or ingame menu), instead of the current two clicks needed. Lastly, the
menu bar is less crowded with only 3 buttons instead of 6.