renderers: make decals work with multiple world scenes
Main view and portals/mirrors call R_CullDecalProjectors, as do multiple world scenes.
A decal not in main view frustum might be visible in portals/mirrors or another scene.
So don't kill decal projectors when culling or reordering them.

Add a new projectorNum variable to decalProjector_t and decal_t.
It's a global instance number so decals aren't projected on the same surface multiple times.

Add a new frameAdded variable to decal_t. It's used so that instant decals aren't reused the same frame.
It's needed so that instant decals aren't projected multiple times sometimes (when the decal was reused).

All decal projectors added this frame when a scene is rendered will be used (including onces before the last scene).