mod: disguiseName replaced by disguisedClientNum in clientInfo_t/player configstring refs #400
This is the first step to get info about uniform owner is in fireteam of
victim @client-side. 
Important: disguisedClientNum itself can't be 0 (because of slot 0)!
and isn't always up2date because game doesn't call
ClientUserinfoChanged() each time covert ops looses disguise (see
PW_OPS_DISGUISED = 0). This shouldn't be an issue - we only use
disguisedClientNum when PW_OPS_DISGUISED is 1.