game: WIP rework the out-of-ammo sound condition event
Previous condition don't allow playing out of ammo event when we run out
of grenade, dynamite, landmine and smoke bomb.
Since dynamite and smoke bomb don't use ammo (see ammoNeeded and "uses"
in weaponTable) we never enter in this condition.
Landmine "uses" field in weapon table was set to 0, which is false.
(Changed) to 1.
So, the only weapons which uses ammo and don't allow no ammo sound are
the grenade and landmine.