game: rework SetWolfSpawnWeapons, AddWeaponToPlayer and AddExtraSpawnAmmo
Add secondary weapon in player classes to be used for checking weapon
validity when spawning. Get equivalent weapon in case the weapon team is
not correct (mp40 <--> thompson, ...).
Remove return value from AddWeaponToPlayer because it is always true and
avoid every check for this return.
Move in AddExtraSpawnAmmo extra *2 clip for Sten and SMG for soldiers.
Tiny optimisation in case weapon don't use ammo so don't need extra.