game: fix firing anim/sound play when charge bar is too low by moving charge time handle to bg side
The best example are mortar and ammo/heal pack :
When pressing continuously the fire button and the charge bar is too low
for firing, a last animation with sound is played without any firing
done.
By moving the charge handle in bg side we ensure the value is updated in
time to check it before a new command is sended by the client with the
old value of the charge time.