Render2: fix the lightmap issue once and for all
Let shaderstage use the rgbGen and alphaGen wich was intended and fix if shader stage references a lightmap, but no lightmap is present (vertex-approximated surfaces), then set cgen to vertex. as this is a colormap we will use generic render for this. this solves best the problem with this issue, Note!
ETL textures on oasis and radar ponds needs tweaking