game: fixed dynamite chain check on secondary objectives, refs #1487
The dyna chaining feature checks if dyna are set on the same objective
to ensure only relevant dyna are exploded in case two objectives are
close to one another.

A condition error was introduced in vanilla, but this was inconsequential
on that codebase. Years later, the onobjective parameter was never set
on secondary objective. The dyna chain feature was however still working,
since the check was comparing 'empty' to 'empty' in G_RadiusDamage()
and etpro_RadiusDamage().

Properly fixing the condition also avoid issues in the dynamite
indicator feature implementation.