R_BuildCubeMaps() more fail safe and efficient.
Also, now rendering only the missing cubemaps (the ones that are not found in an existing cm/cm_xxxx.tga file).
So, if you now suppose to have 5 cm_xxxx files for a map, and you delete cm_0002, then only the cubemaps that are contained in that file cm_0002 are rendered. The rest is loaded from the already existing file(s).

The behavior before was: if the first cm_xxxx file was missing, they all got rendered.
If you missed any other cm_xxxx (than the first), nothing got rendered, but the game would crash (memory copy error).