Pixel Buffer Objects added:
They are buffers for pixeldata.
You can use them to copy data from/to cpu/gpu.
If you issue a pixeltransfer, that pixeldata is not immediately read/written. OpenGL determines when to perform the actual reading/writing.
You must check regularly if the pbo is ready and the pixeldata is ready to be copied.

The render-command to render a cubemap is now using PBOs. (still in a blocking way though)