cgame: fix spawn timer disappearing on pause
This is reverting invalidation spawnTimer stamp when cgs.levelStartTime is changed.
I thought levelStartTime changes only on map restart or map change, but it changes every 500 ms when match is paused too. This is vanilla behaviour which was marked as FIXME in G_RunFrame.
So now timer is enabled once its been set, even after map change, which is not a big deal.
I don't have any idea how to detect map restart or change, but one ugly way around this could be to check whether the levelStartTime change is exactly 500.