cgame: fix SIGSEGV when parsing hud in 64bit arch, refs #1779
I believe structure padding was the culprit. Used same 'trick' that's used in entityStateFields[]. Order of hudComponentFields doesn't matter now.

Not sure whether to use `size_t offset` or `int offset`. `size_t` is supposed to be correct and it should not bring warnings. `int` uses less memory and is big enough now.