client: restore vanilla activatelean behavior, refs #1781
Vanilla 2.60b has +activate leaning enabled by default. After moving the toggle cvar to mod code in fbfd630, this was no longer the case. Restored the code to cl_input to enable it by default. This does not change anything in legacy mod, the mod-sided cvar is still ultimately in control of this feature, this merely enables and restores consistency with 2.60b for +activate lean behavior for mods which do not implement a toggle for this, such as silent mod.