mod: change handling of player events in client game, refs #1957
The issue of doubled sound occurs when client miss predicted events causing them to play again. For example server sometimes predicts 2 events a footstep and fireweapon compared to client only predicting fireweapon and if footstep comes before fireweapon, both will be played.

The proposed solution in PR to this is to ignore all predictable server events (except non-predictable server only generated ones). So in above case we will no longer get a secondary fireweapon event and additional footstep that we miss predicted. However nothing changes for server and nothing changes for how other players are perceived as it's server that decides what events they get, so other players will still hear a footstep and fireweapon.

To do this, client will no longer save events to playerstate_t but to pmoveExt_t and will ignore all predictable events coming from server in playerstate. For prediction we will only check for new non-predictable events and if we find them we will do full prediction as before so we can correctly play them.