client: fix escape key not sending a char event
This readds VET behaviour for escape key. Without it when closing menus like voice chat window (mp_QuickMessage) the UI keycatcher would get unset on KeyUp (instead of char event). This would mean that for a time the BUTTON_TALK would be getting added to user commands as the ui keycatcher is set and because of this all user move commands would get discarded in pmove.