cgame: add spawnpoint visualization
Adds option to visualize spawnpoints in the map with cg_drawSpawnpoints. Displays major spawnpoint name & number, and individual spawnpoints where players spawn. If id key is present in the spawnpoint, it's visualized as well. Only spawnpoints available for your team are drawn. Available in warmup, or if cheats are enabled.