client: don't paste from clipboard using AltGr+V (fixes #2084)
On some layouts, pressing AltGr causes SDL to emit LCTRL and RALT scancodes. This causes problems when trying to type '@', as the clipboard content is pasted first ahead of the intended character.

This is not an optimal fix, but it shouldn't cause any unintentional side-effects.

See: http://forums.libsdl.org/viewtopic.php?t=8590