server: rework server demos, refs #278
The idea with this rework is to remove as much mod specific code from server and make supporting of other mods a possibilty since in current implementation it is not really feasible. This moves the responsibility for mods to add support for it and in lack thereof currently recording and playing will be disabled. The minimum work for server demos recording and playing would be implementing trap calls, G_DemoRunFrame in G_RunFrame and G_DemoStateChanged, which in itself shouldn't be too complicated. In legacy some stuff added currently is optional fe. bot that is requesting stats.

Recording maybe could be enabled for mods without support and during playback use legacy mod server side and correct mod client side (but this will make some stuff like scoreboard unlikely to work, similiar to ETTV with other mods than etpro), but I never got to test this.

It most likely is not the best possible implementation of server demos but I think this was the fastest way to improve them. Watching old server demos will not be possible with this as the code that reads the metadata from beginning of the demo has changed.

I wrote most of the code for fun a year ago and barely touched since so some stuff might not be really finished and I'm not 100% sure it works in every scenario as it was too much for me to test, but at current time I don't know of any critical bugs.