Soft Shadows corrected. The ugly seam at parallax shadow splits is gone. The PCF was sampling outside the split frustum.
Code is improved to sample as few times as needed.

The cvar r_shadowSamples value range has changed. A value of 1 was not working any way.
The current minimum value (2) is too small for greater shadowBlur values, but it could work for small shadowBlur values.
The default shadowSamples is still 3. That does not produce any darkening on splits further away from the camera.