ui: fix potential out of bounds read with uiInfo.map/campaignList
Since UI_LoadArenasFromFile increases the mapcount before it checks for the limits, it will happily increase the count beyond MAX_MAPS limit. While this is sort of stupid by design, it's more convenient to cap it here so the warning about large amount of pk3 files can be displayed easily. This way the user also knows that they have too many maps.