general: Find cjson on Fedora, where cmake config is missing
* Provide simple helper for cJSON location

If BUNDLED_LIBS is disabled and cjson library does not ship its cmake
config for some reason, attempt to find it the other way. That is the
case on current Fedora cjson package.

Provide CJSON_LIBRARY in addition to CJSON_LIBRARIES variable.

* Remove compat CJSON_LIBRARY variable usage

Use just CJSON_LIBRARIES similar to other modules.

* Try cJSON config-only search as the first step

Use config by default even when FindcJSON module is preferred. Try
CONFIG only and if that works, not not attempt other methods. Should
behave exactly the same way as if FindcJSON module were not present.