cgame: fix aspect ratio correction for limbo/debriefing mouse movement
CG_MouseEvent already handles aspect ratio correction for mouse movement in limbo, so this was applying it twice for that. Debriefing cursor handling was done separately from the limbo handling, which was relying on the adjustment done at the syscall, so in order to keep that correct, handle the correction in CG_Debriefing_MouseEvent instead.