game: fix erroneous invalid client print on give command
The give command was always using the first argument to check if it's a valid client, which resulted in commands such as 'give skill' to print an error from the G_ClientNumberFromString function, even though the command was successfully executed. With this change, we check for a valid give command first, and only check for client if the command isn't a valid argument for give.