game: Smooth PM_GroundTrace (OTG) Jump Animation
The kickoff/OffTheGround (OTG) move mechanic has been very oversensitive
when it comes to jump animations, insofar that moving across fairly flat
surfaces with only slight angle changes will cause a kickoff event and
consequently a jump animation event, which causes players to "stumble" /
"hobble" across seemingly flat terrain.

In order to remedy the worst cases, we introduce a directional check in
the kickoff event that gates the jump animation playback.

In order for a kickoff to trigger a jump animation, it now has to be
facing _somewhat_ towards the sky or anyhwere downwards.

The values have been tuned to still cause a jump animation when getting
a kickoff via prominent terrain features.

The gameplay mechanical code of the kickoff events remains completely
unchanged, as such it is a purely cosmetic.