cgame: Fix 'FIGHT!' voice line when following
Once warmup ends while following another player as spectator, the
"FIGHT!" line gets written but the voice line is not played.

The issue happens as 'listener_number' in 'src/client/snd_dma.c' is
still set to the entity of the player to be followed, despite that after
warmup a spectator is reset to unfollow and due to timing,
'listener_number' is not correctly reset to the spectator (via
'respatialize') before the voice line sample is actually queued.

Thus the voice line is played in the wrong place and can't be heard by
the spectator.

This commit fixes this issue, not by addressing the timing issue between
'respatialize' and match start, but by always forcing the "FIGHT" voice
line to be played on the spectator's position when following another
player.

Fixes https://github.com/etlegacy/etlegacy/issues/522