mod: change the way /scores server command works
- add back /scores and auto updates to scores in tvgame
- does not break old demos

Note:

/scores command causes lag client side, because it is really big it fragments the packet (mtu is 1300) if sent all at once (just 10 players can cause the packet to be split even 3 times). This causes the player to get rate limited which in turn means delayed packets which means players has nothing to interpolate to.

Instead of sending everything at once the strings are saved per player basis in qagame and added to snapshot one per clientthink, this still can be too much on some settings (com_maxfps and sv_fps) and with many players on server.
Command is displayed on client only after receving all strings. The downside to this is there will be more delay when the scores commands appears on client.

Pretty much the only reason I would like to add it is to be able to do automatic updates of scores for etltv (only legacy mod) without causing server lag every X seconds (etltv server is also a client that gets it's packets fragmented).