cgame: Trim last 2 flame sprites
The 2 frame sprites fade out the flame sprite (additionally to the alpha
channel whichis already being faded out in code).

This however leads to situations where you think the flame chunk just
disappeared but it is actually still there.

This is significantly reduced by trimming the 2 last frames of a flame
chunk sprite, which spreads the remaining frames out over the same
duration.