ui: fix replay menu path resetting when re-opening the menu
The first byte of uiInfo.demos.path was getting reset whenever LoadDemos was called, which caused the current display path to always be blank when opening replay menu. This caused the list to appear blank if the list was scrolled to an entry which was not visible on screen at the time the menu was exited. The current path is now always remembered correctly, and the replay menu will open back to the folder with previous entry selected when re-opening.