ui: fix replay menu path resetting when re-opening the menu, fixe #2525
The first byte of uiInfo.demos.path was getting reset whenever LoadDemos was called, which caused the current display path to always be blank when opening replay menu. This caused the list to appear blank if the list was scrolled to an entry which was not visible on screen at the time the menu was exited. The current path is now always remembered correctly, and the replay menu will open back to the folder with previous entry selected when re-opening.

It seems this was the intended behavior when the folder support was created, as there's code in UI_LoadDemos which suggests that it was designed to remember the last opened path. Whether this is desired or if the menu should always reset back to the root demos folder is up to discussion I guess.

609fdca