renderer: set uploadWidth/Height for internal images, fixes #2250
`image->uploadWidth/Height` were not set for internal images, which caused `imagelist` command to report incorrect texel cound and possibly over/underflow. This happened on every renderer init with the call to `R_CreateImage` from `R_InitGamma`, which caused there to always be an image in the list with unset `uploadWidth/Height` (unless `r_ignorehwgamma` was set).