etmain: Reframe revive 3P animation
Fit the revive animation (i.e. standup) into the gameplay relevant
duration of the player standing up.

Instead of having it scale throughout the whole revive-shield duration,
towards the end of which the player is actually able to move.

This fixes hovering and gives a visuall hint to discern a player
standing up after being revived without moving, and being allowed to
move while still the revive shield is up.