client: Rely soley on 3P anim priority system (#2781)
* client: Rely soley on 3P anim priority system

Priority values for 3P animations currently co-exist with passing the
`force` parameter in-code, which is manually set in various code-paths
for various reasons.

This commit completely disables forcing 3P animations via code and
instead soley relies on the priority values that are set in the '.aninc'
files.

This greatly simplifies anim prioritizations, by making the priority
value the de facto single source of truth.

Further changes are introduced due-to this:

- Switching weapons during a RAISE animation will now reset the RAISE
  animation properly (previously switching quickly between weapons would
  keep the currently running RAISE animation).

- Knife attacks now loop more smoothly.

- Attaching rifle grenades is now not canceled via ladder movement.

- Keep 'nopower' animation when jumping, instead of being broken by it.

And further changes are introduced alongside:

- Animations with the same priority value now override each other.
  RELOAD and FIRING now share the same priority value as a consequence.

- Raising a knife while proned now does not play the
  'prone_knife_attack' animation.

- Pliers now only play the firing animation as long as the attack button
  is held.

- Reframe both 'raise_knife' and 'firing_knife'.

* game: Remove deprecated 3P animation 'force' param

* etmain: Reframe revive 3P animation

Fit the revive animation (i.e. standup) into the gameplay relevant
duration of the player standing up.

Instead of having it scale throughout the whole revive-shield duration,
towards the end of which the player is actually able to move.

This fixes hovering and gives a visuall hint to discern a player
standing up after being revived without moving, and being allowed to
move while still the revive shield is up.