client: add cvar cl_interpolation
- increase MAX_CLIENT_MARKERS from 17 to 40 (antilag buffer, supports 400ms ping (from 170ms) on sv_fps 100)

Note about cl_interpolation:
Buffers server packets to smooth over packetloss/ping instability.
Values 0-4 depending on 'sv_fps' and 'snaps' (scales off the smaller value of the two).
Set to 0 for most responsive gameplay.

Similiar to other engines the cvar allows to change the intepolation amount to smoother out the gameplay at the cost of responsivnes.
Interpolation in ET engine by default is tied to sv_fps/snaps, meaning at sv_fps 20 interpolation is locked to 50ms, at sv_fps 100 interpolation is locked to 10ms.
I believe this turned out to be a problem few years ago when I tried to push for sv_fps 100 as some players reported that everyone is laggy. While I didn't have any information past that my assumption was and is that interpolation is just too small.
sv_fps 100 means that a client expects new snapshot every 10ms. If a player uses com_maxfps 125 (8 ms frametime) that means that it basically expects a new snapshot every frame which could be very unreliable when not in a LAN setting but over internet.
cl_interpolation is meant to help with that, by buffering additional snapshot(s) before rendering them and extending the window where a delayed/missing snapshot becomes an issue.
Because as far as I know noone ever played ET on sv_fps 10, the default max interpolation is still 50ms, meaning on sv_fps 20 the cvar has no effect.
On sv_fps 40 only allowed values are 0 or 1, which means that at most one additional snapshot can be buffered, the result of that is it will feel as if playing on sv_fps 20 (in terms of noticing a respones for example a hitsound).
sv_fps 100 supports values from 0 to 4, 4 would mean as if playing on sv_fps 20, value 1 would mean intepolation of 20ms so faster than sv_fps 40 by 5ms, while not much it could be good enough for people that internet connection would not allow for enjoyable experience on sv_fps 100 with default 10ms interpolation.