client: add possibility to pause sounds (SDL only)
- add engine extension CG_MATCHPAUSED to let engine know when match is paused
- add SoundsPause function to soundInterface_t which is called on CG_MATCHPAUSED
- add reload sound as pausable sound (maybe more in the future)
- add cvar s_debugPause (CVAR_CHEAT) for easier debugging

Note:
Adding pausable sounds is possible during the time when sounds are paused, which makes them play only after they are unpaused.
This could be changed in cgame to not allow adding such sounds (by for example removing the flag before starting the sound), right now the developer needs to be aware of it and not add them.
Looping sounds and some other sounds not supported because some engine calls don't support setting sound flags.