game: fix oob write to dmgReceivedSts when attacker wasn't a client
attacker->s.number can be any entity in the map, which meant that this could easily cause an out of bounds write. This would manifest into this write getting written into the next gclient_t struct in gentity_t struct, which wrote garbage data to the next players playerstate most likely (as that's the first field in gclient_t struct).