cgame: Add cg_muzzleFlashScale
Muzzle Flashes are known to be a visibility problem for a long time,
which is why 'cg_muzzleFlash' was added to disable them.

This decision however might've been overhasty, as gradual reduction of
muzzle flash size seems to alleviate most of the problems without
removing parts of the game altogether.

This commit adds a new 'cg_muzzleFlashScale' CVAR and after extensive
experimentation, defaults it to '0.8', which provides a better
visibility default, without a generally noticable shrink of muzzle flash
size.