mod: fix uiMenuCommand_t enum order back to normal
Years ago there were new menu commands added at the middle of the uiMenuCommand_t enum. This caused issues with other mods running on ETL, as the enum is also compiled into the engine, which sends auto-update popups via VM_Call to UI. Because UIMENU_WM_AUTOUPDATE in legacys bg_public.h was different from what the menu order is in etmain and derivative mods, this would break the auto-update popups on other mods, and potentially open completely wrong menus in the UI when the game is started and updates are checked (for example, ETJump opens a completely unrelated popup menu when the game is started due to UIMENU_WM_AUTOUPDATE having value 16 instead of 14 in engine).

This restores the original enum order for entries 0-15, and moves the legacy-exclusive enum entries to the end of the enum, like they should have been inserted to begin with.