etmain: Tweak airplane visuals
Modified textures with added pseudo-shadows and some grit.

Also removed the transparency effects, as the models caused transparency
overlap between their own faces and there doesn't seem to be a need for
it atm. - it however should be fixed and re-enabled at a later point,
but this is not possible without recasting the model geom and doing the
UV wrap again.

In particular, the issue we have is that we failed to have solid
airplanes that become transparent only after a certain time - either
they would always be transparent and we could fade them more - or they
were always solid - we chose the latter for now, but you could probably
have both at a later point.