cgame: Limbo defaults to SMG if Soldier can
Currently the SMG will never be autoselected when switching to the
Soldier class through Limbo menu.

The reason for that is that 'SetDefaultWeapon(SECONDARY_SLOT)' is called
once the Class is changed and never thereafter - however, when the Class
is changed, the default primary weapon is set to the SMG and the
consequent 'SetDefaultWeapon(SECONDARY_SLOT)' call will reject the SMG,
as it always rejects weapons that are already set in the primary slot.

To circumvent this we now track if a user ever switched the weapon panel
to the secondary panel (which will reset if they change either class or
team) and if they never have opened it, we call
'SetDefaultWeapon(SECONDARY_SLOT) **again** right before we actually
place the `team` command in 'CG_LimboPanel_SendSetupMsg'.

The end effect is that just switching to Soldier and selecting a
heavyweapon will now always prioritize the SMG as long as the SMG perk
is available.

Also adds 'z' as an alias for confirming limbo selection (think QWERTY
keyboards et al.)