tvgame: fix animation files breaking mod loading on different mods than legacy
BG animation parser was failing to load different mods animation files because of various changes that have been made in legacy mod.
Animation parsing is not required, but I've had to put it back in some time ago because of some weird random crashes in pmove.
This time relevant pmove code was extracted from bg_pmove and most bg_ files are no longer needed.
That also made it possible to fix various pmove prediction issues on mods like ETJump or ETPro, which are incompatible with Legacy by default.