cgame: Fix noAmmoSound on multiple grenade bank presses
Using the grenade bank key more than once would make a noAmmoSound as it
tried to switch to the other team's grenade - which of course would be
unavailable :')

Also restrict the unavailable noAmmoSound only to your team's grenade,
as grenades are the only weapon who seem to become unavailable when
being used up anyhow.

Also fix a case where multiple satchel bank presses would cause the
noAmmoSound to be played as it tried to switch to the satchel det, which
defines ammo, yet has 'useAmmo' == qfalse.