game: Tweak Spectator aim-following
Currently, as a spectator, issueing an `ATTACK` cmd while hovering over
a player will move the view to that particular player.

However, if the trace (which determines if a spectator is currently
hovering over another player) misses, the "next player" will be swapped
to instead - as would issueing an `ATTACK` cmd without hovering any
player by default.

This constitutes an issue, whereby a spectator who intends to follow a
particular player, misses his aim and instead to switching to the
intended player, switches to a seemingly random player.

To give an explicit solution for this intention, holding down the
`SPRINT` key while issueing an `ATTACK` cmd will now never move the view
to the "next player" - so missing a player while doing so will simply do
nothing in this instance, giving the spectator the opportunity for
another attempt.

Previously the `SPRINT` & `ATTACK` combination was used to differentiate
cycling between non-bot and bot players, this combination is hereby
moved to `WALKING` & `ATTACK` instead.

Furthermore, the way by which we determine which player is being aimed
is improved in the following way:

Instead of making a single "fat" trace (with a +/- 64 unit beam), we
make 4 separate traces for each `ATTACK` cmd while spectating.

The first one is a simple unsized hitscan, which checks if the spectator
has directly hit the player - if not, increasing beam-widths are tried
in increasing size to narrow down the potential target.

The maximum beam-width tried however has remained the same (+/- 64 units).

This commit therefore attempts to greatly increase the intuitiveness of
using this particular feature.