cgame: Unique names when cloning via HUD Editor
Previously cloning a HUD would just use the parent HUD name and suffix
with `%s_copy`, no matter if it existed prior or not, creating multiple
HUDs with the same name.

Instead of doing this, we now check if `%s_copy` is free, if it's not,
we try `%s_copy%d` >= 2 && < 100 until we find a free spot and use that
as a new name.

If no free spot can be found, we print an error and stop.