cgame: Add 'Toggle Help' Button to HUD Editor
Currently the Help Text always gets shown when opening the HUD Editor
for the first time since the game was started.

This makes sure that users take note of it, however it also becomes
cumbersome after you know about it's existance.

Instead of showing the help text always at the start, this commit adds a
`Toggle Help` Button and gives it a special color to make it easily
noticeable.