cgame: Return to use `.tga` crosshairs, instead of `.svg`
As we've received reports of issues with the SVG crosshairs (#2942),
this change stops using `.svg` crosshairs by default .

Also adds a 'cg_crosshairSVG' CVAR (default 0), which allows opting in
to the `.svg` crosshairs by setting it to 1.

As currently paths in the renderer always prioritize `.svg` files over
`.tga` files, no matter if we specify a file extension explicitly or
not, we have to give the `.svg` crosshairs unique file names when
stripping the file extension.