qcommon: allow viewlog to be spawned during runtime
For reasons unknown, this was made impossible 10 years ago in af0ae3bac42610c028dea4eea6fed0333d50c9f4 by removing the check for the cvar modification from Com_Frame, which made it impossible to spawn the console any other way besides as a startup option. The viewlog remains a cheat cvar though, so it's mainly useful for debugging things and specifically in ETJump as it provides a cvar unlocker to view the console window.