Refactored the math code quite a bit. Axis and mat9 are now separated cleanly
idTech axis fields are row major 9 slot rotation matrices, while the
mat16 was column based and used with OpenGL (which is column major by
default like glm for example).

Now Axis is the "old" style 3x3 float matrix that is row major and the
others are single level arrays with 9 or 16 slots respectively, and they
are column major.