Remove common graphic settings restriction.
While there have been a few discussions on this,
most notably: https://github.com/etlegacy/etlegacy/discussions/1818

Enforcing arbitrary hardlimits in engine is not
the right place.

Most especially given the recent developement
of more and more people reaching for custom
clients to bypass these.

While we may choose to enforce some limits via
sv_cvar it is still ultimately the community's
and server owners' choice which values to enforce.

This is the sole point and purpose of providing
sv_cvar restrictions in first place, to not have
to rely on hardcoded limits that can be
circumvented by anyone with a litte bit of
know-how.

ETLegacy still provides custom configs and may
suggest which limits should be enforced or not.
These are not cheat protected cvars and we're not
treating them as such either.

There's a clear distinction to be made.