Add version info to development builds loading panel (#2998)
* Add version info to development builds loading panel

When the `DEVELOPMENT BUILD` label was added, `ETLEGACY_VERSION` was removed due to space constraints.  The version information is extremely useful to have so this commit aims to solve both the need for the development build label as well as having version information on the loading panel.

Positioning of the start of the print was adjusted to account for having two lines versus one, as well as text sizing adjustments to ensure two lines are still readable within the desired area.  Keeping the version on a separate lines also ensures space for longer build strings should they be necessary over the current example.