cgame: fix incorrect player position at the start of intermission
Don't run full limbopanel setup when client info is updated, only run the relevant part (weapons), otherwise we get placed to a wrong position when intermission starts and potentially break any intermission scenes involving entities, because they are no longer in our PVS.

This also fixes 'getspawnpt' being run immediately when intermission starts, similar to what 'obj' did, which prompted the change to allow it during intermission (which broke this whole thing). We probably should allow it to run during intermission though, as limbo panel can be opened during intermission to switch teams, which will trigger the prompt that the command is not allowed during intermission.

Opening the limbo panel during intermission will still predictably break the intermission screen (as it triggers sending 'obj -1', which is what broke this), but fixing that is out of scope for this commit.