easybuild.bat: as starting point add ability to create another mods
Rest of functionality needs to be ported from old existing source code for specific mod, which is simply too painfull to work with on current new operating systems/compilers.

Those functionalities needs to be in seperated  repos/forked repos or in another branches.

Same changes comming in easybuild.sh file